fix: changes after review
