+++ package.json
@@ -16,9 +16,7 @@
     "@mui/x-date-pickers": "^7.23.6",
     "axios": "^1.7.9",
     "i18next": "^24.2.1",
-    "i18next-browser-languagedetector": "^8.0.2",
     "i18next-chained-backend": "^4.6.2",
-    "i18next-http-backend": "^3.0.1",
     "i18next-resources-to-backend": "^1.2.1",
     "joi": "^17.13.3",
     "jwt-decode": "^4.0.0",
@@ -29,7 +27,6 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-i18next": "^15.4.0",
-    "tailwindcss-animated": "^1.1.2",
     "uuid": "^11.0.5"
   },
   "devDependencies": {
