+++ src/app/page.tsx
@@ -13,24 +13,20 @@ export const metadata: Metadata = {
 
 export default function Home() {
     return (
-      <>
-          <LocalizationProvider>
-              <SnackbarProvider>
-                  <SessionProvider>
-                      <AuthProvider>
-                          <NavigationTopBar />
-                          <div
-                              className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
-                              <main className="flex flex-col gap-8 row-start-2 items-center sm:items-start">
-                                  <div className="flex gap-4 items-center flex-col sm:flex-row">
-                                      <InstantMeetingButton />
-                                  </div>
-                              </main>
+      <LocalizationProvider>
+          <SnackbarProvider>
+              <SessionProvider>
+                  <AuthProvider>
+                      <NavigationTopBar />
+                      <main
+                          className="flex flex-col h-screen max-h-[80vh] items-center justify-center m-4 font-[family-name:var(--font-geist-sans)]">
+                          <div className="flex flex-1 items-center justify-center m-4">
+                              <InstantMeetingButton />
                           </div>
-                      </AuthProvider>
-                  </SessionProvider>
-              </SnackbarProvider>
-          </LocalizationProvider>
-      </>
+                      </main>
+                  </AuthProvider>
+              </SessionProvider>
+          </SnackbarProvider>
+      </LocalizationProvider>
   );
 }
